Converting floating-point arithmetic to fixed-point arithmetic remains an unresolved problem.

+++ program/algorithm/transform.h
@@ -5,6 +5,8 @@
 #ifndef MINIFOC_ALGORITHM_TRANSFORM_H_
 #define MINIFOC_ALGORITHM_TRANSFORM_H_
 
-void foc_calculate_dutycycle(float elect_angle, float d, float q, float* u, float* v, float* w);
+#include "stdint.h"
+
+void foc_calculate_dutycycle(int32_t elect_angle, int32_t d, int32_t q, int32_t* u, int32_t* v, int32_t* w);
 
 #endif  //MINIFOC_ALGORITHM_TRANSFORM_H_
